refactor WorkRepository to WorkRequests
