--- app/src/main/java/com/github/sergereinov/aa2020/worker/WorkRepository.kt
@@ -1,24 +1,0 @@
-package com.github.sergereinov.aa2020.worker
-
-import androidx.work.*
-import java.util.concurrent.TimeUnit
-
-class WorkRepository {
-    companion object {
-        val refreshMoviesNow
-            get() =
-                OneTimeWorkRequest.Builder(RefreshMoviesWorker::class.java).build()
-
-        val refreshMoviesPeriodic: PeriodicWorkRequest
-            get() {
-                val constraints = Constraints.Builder()
-                    .setRequiredNetworkType(NetworkType.CONNECTED)
-                    .setRequiresCharging(true)
-                    .build()
-                return PeriodicWorkRequest
-                    .Builder(RefreshMoviesWorker::class.java, 8, TimeUnit.HOURS)
-                    .setConstraints(constraints)
-                    .build()
-            }
-    }
-}
